movie/requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11.3
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.0
django-guardian==1.4.9
django-import-export==1.0.0
django-pure-pagination==0.3.0
django-reversion==2.0.13
django-simple-captcha==0.5.5
djangorestframework==3.6.3
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.3
idna==2.6
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
MySQL-python==1.2.5
odfpy==1.3.6
openpyxl==2.5.3
Pillow==5.1.0
pkg-resources==0.0.0
PyMySQL==0.8.1
pytz==2018.4
PyYAML==3.12
requests==2.18.4
shortuuid==0.5.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
XlsxWriter==1.0.4
xlwt==1.3.0
